1.0.7
Fixed position not being set by shared settings
Added support for API changes in 9.0
Added support for seding settings when a party is formed or changed.  Settings will only be sent from the party leader.
Disabled sharing settings if you are not in a party, this prevents unnecessary addon messages